+++ agents-flex-store/agents-flex-store-qdrant/src/test/java/com/agentsflex/store/qdrant/QdrantVectorStoreTest.java
@@ -20,17 +20,17 @@ public class QdrantVectorStoreTest {
         Document doc1 = new Document();
         doc1.setId(1L);
         doc1.setContent("test1");
-        doc1.setVector(new double[]{5.2, 4.4});
+        doc1.setVector(new float[]{5.2f, 4.4f});
         list.add(doc1);
         Document doc2 = new Document();
         doc2.setId(2L);
         doc2.setContent("test2");
-        doc2.setVector(new double[]{5.2, 3.9});
+        doc2.setVector(new float[]{5.2f, 3.9f});
         list.add(doc2);
         Document doc3 = new Document();
         doc3.setId(3);
         doc3.setContent("test3");
-        doc3.setVector(new double[]{4.9, 3.4});
+        doc3.setVector(new float[]{4.9f, 3.4f});
         list.add(doc3);
         db.store(list, options);
     }
@@ -41,7 +41,7 @@ public class QdrantVectorStoreTest {
         StoreOptions options = new StoreOptions();
         options.setCollectionName("test_collection1");
         SearchWrapper search = new SearchWrapper();
-        search.setVector(new double[]{5.2, 3.9});
+        search.setVector(new float[]{5.2f, 3.9f});
         //search.setText("test1");
         search.setMaxResults(1);
         List<Document> record = db.search(search, options);
